* commit '7e14be1fbfcd2922b2a7b74e52018dfcb74d9b0b':
  Added database interface that describes the access to objects by index getObjectDataForIndex()
  Small typing and doc comment improvements
  Changed the Database test to use the database based test case
  Fixed a bug with Indexes when constructing a Database with data
  Renamed getObjectDataForIndex() to getObjectDataForIndexIfSet() Utilize Database traits
  Added trait for Database state
  Added trait for Indexable Databases
  Fixed a bug where a Handler Result was constructed incorrect
  Added a check for the Document collection passed to the Expand Resolver
  Added a check if a Handler implements a special Handler Action
  Formatted code
  Added additional check before calling setObjectDataForIndex() with null argument
  Added additional check in remove command
  Updated the homepage URL in the composer.json
  Updated the composer.json
  Merged composer.lock changes

+++ Classes/Console/Data/RemoveCommand.php
@@ -49,15 +49,10 @@ class RemoveCommand extends AbstractDataCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-//		$databaseIdentifier = $input->getArgument('database');
-//		$objectIdentifier = $input->getArgument('identifier');
-//		$database = $this->coordinator->getDatabase($databaseIdentifier);
-//		$document = $database->findByIdentifier($objectIdentifier);
-//		if (!$document) {
-//			throw new InvalidDataException(sprintf('Object with ID "%s" not found in database %s', $objectIdentifier, $databaseIdentifier));
-//		}
-
         $document = $this->findDataInstanceFromInput($input);
+        if ($document === null) {
+            return;
+        }
         $database = $this->findDatabaseInstanceFromInput($input);
         $database->remove($document);
         $objectIdentifier = $document->getId();
